--- blns/blns.py
@@ -1,155 +1,0 @@
-# -*- coding: utf-8 -*-
-import json
-import random
-
-blns_list = [
-    "",
-    "undefined",
-    "null",
-    "NULL",
-    "nil",
-    "NIL",
-    "0",
-    "1",
-    "1.00",
-    "$1.00",
-    "1/2",
-    "1E2",
-    "1E02",
-    "1E+02",
-    "-1",
-    "-1.00",
-    "-$1.00",
-    "-1/2",
-    "-1E2",
-    "-1E02",
-    "-1E+02",
-    "1/0",
-    "0/0",
-    "0.00",
-    "0..0",
-    ".",
-    "0.0.0",
-    "0,00",
-    "0,,0",
-    ",",
-    "0,0,0",
-    "--1",
-    "-",
-    "-.",
-    "-,",
-    "999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999",
-    "NaN",
-    "Infinity",
-    "-Infinity",
-    ",./;'[]\\-=",
-    "<>?:\"{}|_+",
-    "!@#$%^&*()`",
-    "Ω≈ç√∫˜µ≤≥÷",
-    "åß∂ƒ©˙∆˚¬…æ",
-    "œ∑´®†¥¨ˆøπ“‘",
-    "¡™£¢∞§¶•ªº–≠",
-    "¸˛Ç◊ı˜Â¯˘¿",
-    "ÅÍÎÏ˝ÓÔÒÚÆ☃",
-    "Œ„´‰ˇÁ¨ˆØ∏”’",
-    "`⁄€‹›ﬁﬂ‡°·‚—±",
-    "⁰⁴⁵",
-    "₀₁₂",
-    "⁰⁴⁵₀₁₂",
-    "'",
-    "\"",
-    "''",
-    "\"\"",
-    "'\"'",
-    "\"''''\"'\"",
-    "\"'\"'\"''''\"",
-    "田中さんにあげて下さい",
-    "パーティーへ行かないか",
-    "和製漢語",
-    "部落格",
-    "사회과학원 어학연구소",
-    "社會科學院語學研究所",
-    "울란바토르",
-    "𠜎𠜱𠝹𠱓𠱸𠲖𠳏",
-    "ヽ༼ຈل͜ຈ༽ﾉ ヽ༼ຈل͜ຈ༽ﾉ ",
-    "(｡◕ ∀ ◕｡)",
-    "｀ｨ(´∀｀∩",
-    "__ﾛ(,_,*)",
-    "・(￣∀￣)・:*:",
-    "ﾟ･✿ヾ╲(｡◕‿◕｡)╱✿･ﾟ",
-    ",。・:*:・゜’( ☻ ω ☻ )。・:*:・゜’",
-    "(╯°□°）╯︵ ┻━┻)  ",
-    "(ﾉಥ益ಥ）ﾉ﻿ ┻━┻",
-    "😍",
-    "👩🏽",
-    "👾 🙇 💁 🙅 🙆 🙋 🙎 🙍 ",
-    "🐵 🙈 🙉 🙊",
-    "❤️ 💔 💌 💕 💞 💓 💗 💖 💘 💝 💟 💜 💛 💚 💙",
-    "✋🏿 💪🏿 👐🏿 🙌🏿 👏🏿 🙏🏿",
-    "🚾 🆒 🆓 🆕 🆖 🆗 🆙 🏧",
-    "0️⃣ 1️⃣ 2️⃣ 3️⃣ 4️⃣ 5️⃣ 6️⃣ 7️⃣ 8️⃣ 9️⃣ 🔟",
-    "１２３",
-    "١٢٣",
-    "ثم نفس سقطت وبالتحديد،, جزيرتي باستخدام أن دنو. إذ هنا؟ الستار وتنصيب كان. أهّل ايطاليا، بريطانيا-فرنسا قد أخذ. سليمان، إتفاقية بين ما, يذكر الحدود أي بعد, معاملة بولندا، الإطلاق عل إيو.",
-    "בְּרֵאשִׁית, בָּרָא אֱלֹהִים, אֵת הַשָּׁמַיִם, וְאֵת הָאָרֶץ",
-    "הָיְתָהtestالصفحات التّحول",
-    "​",
-    " ",
-    "᠎",
-    "　",
-    "﻿",
-    "␣",
-    "␢",
-    "␡",
-    "‪‪test‪",
-    "‫test‫",
-    " test ",
-    "test⁠test‫",
-    "⁦test⁧",
-    "test\u0000",
-    "test\u001b",
-    "test",
-    "test test test",
-    "testကtest",
-    "test﷐",
-    "Ṱ̺̺̕o͞ ̷i̲̬͇̪͙n̝̗͕v̟̜̘̦͟o̶̙̰̠kè͚̮̺̪̹̱̤ ̖t̝͕̳̣̻̪͞h̼͓̲̦̳̘̲e͇̣̰̦̬͎ ̢̼̻̱̘h͚͎͙̜̣̲ͅi̦̲̣̰̤v̻͍e̺̭̳̪̰-m̢iͅn̖̺̞̲̯̰d̵̼̟͙̩̼̘̳ ̞̥̱̳̭r̛̗̘e͙p͠r̼̞̻̭̗e̺̠̣͟s̘͇̳͍̝͉e͉̥̯̞̲͚̬͜ǹ̬͎͎̟̖͇̤t͍̬̤͓̼̭͘ͅi̪̱n͠g̴͉ ͏͉ͅc̬̟h͡a̫̻̯͘o̫̟̖͍̙̝͉s̗̦̲.̨̹͈̣",
-    "̡͓̞ͅI̗̘̦͝n͇͇͙v̮̫ok̲̫̙͈i̖͙̭̹̠̞n̡̻̮̣̺g̲͈͙̭͙̬͎ ̰t͔̦h̞̲e̢̤ ͍̬̲͖f̴̘͕̣è͖ẹ̥̩l͖͔͚i͓͚̦͠n͖͍̗͓̳̮g͍ ̨o͚̪͡f̘̣̬ ̖̘͖̟͙̮c҉͔̫͖͓͇͖ͅh̵̤̣͚͔á̗̼͕ͅo̼̣̥s̱͈̺̖̦̻͢.̛̖̞̠̫̰",
-    "̗̺͖̹̯͓Ṯ̤͍̥͇͈h̲́e͏͓̼̗̙̼̣͔ ͇̜̱̠͓͍ͅN͕͠e̗̱z̘̝̜̺͙p̤̺̹͍̯͚e̠̻̠͜r̨̤͍̺̖͔̖̖d̠̟̭̬̝͟i̦͖̩͓͔̤a̠̗̬͉̙n͚͜ ̻̞̰͚ͅh̵͉i̳̞v̢͇ḙ͎͟-҉̭̩̼͔m̤̭̫i͕͇̝̦n̗͙ḍ̟ ̯̲͕͞ǫ̟̯̰̲͙̻̝f ̪̰̰̗̖̭̘͘c̦͍̲̞͍̩̙ḥ͚a̮͎̟̙͜ơ̩̹͎s̤.̝̝ ҉Z̡̖̜͖̰̣͉̜a͖̰͙̬͡l̲̫̳͍̩g̡̟̼̱͚̞̬ͅo̗͜.̟",
-    "̦H̬̤̗̤͝e͜ ̜̥̝̻͍̟́w̕h̖̯͓o̝͙̖͎̱̮ ҉̺̙̞̟͈W̷̼̭a̺̪͍į͈͕̭͙̯̜t̶̼̮s̘͙͖̕ ̠̫̠B̻͍͙͉̳ͅe̵h̵̬͇̫͙i̹͓̳̳̮͎̫̕n͟d̴̪̜̖ ̰͉̩͇͙̲͞ͅT͖̼͓̪͢h͏͓̮̻e̬̝̟ͅ ̤̹̝W͙̞̝͔͇͝ͅa͏͓͔̹̼̣l̴͔̰̤̟͔ḽ̫.͕",
-    "Z̮̞̠͙͔ͅḀ̗̞͈̻̗Ḷ͙͎̯̹̞͓G̻O̭̗̮",
-    "˙ɐnbᴉlɐ ɐuƃɐɯ ǝɹolop ʇǝ ǝɹoqɐl ʇn ʇunpᴉpᴉɔuᴉ ɹodɯǝʇ poɯsnᴉǝ op pǝs 'ʇᴉlǝ ƃuᴉɔsᴉdᴉpɐ ɹnʇǝʇɔǝsuoɔ 'ʇǝɯɐ ʇᴉs ɹolop ɯnsdᴉ ɯǝɹo˥",
-    "00˙Ɩ$-",
-    "<script>alert('XSS')</script>",
-    "<img src=x onerror=alert('XSS') />",
-    "<svg><script>0<1>alert('XSS')</script> ",
-    "\"><script>alert(document.title)</script>",
-    "'><script>alert(document.title)</script>",
-    "><script>alert(document.title)</script>",
-    "</script><script>alert(document.title)</script>",
-    "< / script >< script >alert(document.title)< / script >",
-    " onfocus=alert(document.title) autofocus ",
-    "\" onfocus=alert(document.title) autofocus ",
-    "' onfocus=alert(document.title) autofocus ",
-    "＜script＞alert(document.title)＜/script＞",
-    "<sc<script>ript>alert('XSS')</sc</script>ript>",
-    "1;DROP TABLE users",
-    "1'; DROP TABLE users--",
-    "/dev/null; touch /tmp/blns.fail ; echo",
-    "../../../../../../../../../../../etc/passwd%00",
-    "../../../../../../../../../../../etc/hosts",
-    "() { 0; }; touch /tmp/blns.shellshock1.fail;",
-    "() { _; } >_[$($())] { touch /tmp/blns.shellshock2.fail; }"
-]
-
-
-def all():
-    """Get all blns in an array"""
-    return blns_list
-
-
-def one(min_len=0):
-    """One Random String"""
-    while(True):
-        nl = random.choice(all())
-        if(nl.__len__() >= min_len):
-            return nl
